src/data/requirements.txt
rasterio
pandas
requests
sentinelsat
baitsss
wslib